application_backend/requirements.txt
Django==2.2
djangorestframework==3.9.2
psycopg2==2.7.5
markdown 
django-filter
